added logic into decode and execute loop:

+++ include/core.h
@@ -1,6 +1,8 @@
+#include <stdint.h>
 
-int decript(int data_for_decript);
-void do_instruct(int instruction);
+
+uint16_t decript(int data_for_decript);
+void do_instruct(uint16_t instruction);
 void decrease_timers();
 void move_pointer_by_next_instruct();
 void update_system_condition();
